config : Add flag to indicate the data path is file or directory
The changes implements a boolean flag as a data member to the structure
DataSyncConfig which will indicate whether each data path configured in
the JSON config files is a file or directory.

This flag is useful to check the data type during data-sync operations
especially when the configured data path not exists in the filesystem.

Change-Id: I1d53c123293b13a9b05fae008a6693dd35ffa762
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>