inotify: pass the necessary arguments explicitly
This update adds flexibility to the inotify framework by allowing the
path to be watched, include list, and exclude list to be passed
separately with include, exclude list as optional.

Previously, the framework stored a reference to the data-sync
configuration object and relied on it to obtain the watch path and
include/exclude lists. While this reduces argument passing and copying
during watcher construction, it limits the reusability of the inotify
component in scenarios where a data-sync configuration object is not
relevant.

A practical example is sibling notification handling, where we need to
monitor a directory for notification requests — independent of any
data-sync configuration.

Tested:

Verified in simics that no existing functionality is breaking with
this change.
    - Able to add watches for the paths in the data-sync config
    - Processing of exclude and include list works fine.

Change-Id: I370e0b127c714ca0a82fd1a4fc166ac6b91c5492
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>